+++ frontend/src/index.css
@@ -17,7 +17,7 @@
 		@apply text-white bg-[#E87030] hover:bg-[#e87030]/80 focus:outline-none focus:ring-4 focus:ring-[#e87030]/30 font-medium rounded-full text-sm px-5 py-2.5 text-center me-2 mb-2 border-none;
 	}
 	.signup-form{
-		@apply form-control gap-3 p-6 md:px-10 md:py-6 w-full max-w-lg text-white border bg-custom-bg-dark bg-opacity-80 rounded-xl mt-16 shadow-lg;
+		@apply form-control gap-3 p-6 md:px-10 md:py-6 w-full max-w-lg text-white border bg-custom-bg-dark bg-opacity-80 rounded-xl mt-16 shadow-lg backdrop-blur-sm;
 	}
 	.signup-input{
 		@apply input input-bordered border-white rounded-full bg-opacity-60;
